feat: create a leadership card
