+++ src/components/LeaderProfileCard/LeaderProfileCard.module.css
@@ -1,0 +1,62 @@
+.card {
+    --card-bg: #ECF9F8;
+
+    .textContainer {
+        display: flex;
+        flex-direction: column;
+        justify-content: space-between;
+        height: 100%;
+        padding: 27px 30px;
+        gap: 10px;
+        background-color: var(--card-bg);
+    }
+
+    .textDecorationNone {
+        color: inherit;
+        text-decoration: none;
+        cursor: pointer;
+        align-self: center;
+    }
+
+    .button {
+        font-size: 16px;
+        font-weight: 600;
+        background-color: #F4F9FE;
+
+        &:hover {
+            background: #007ebb;
+            color: #fff;
+        }
+    }
+
+    .heading {
+        padding-bottom: 16px;
+
+        &3 {
+            font-size: 28px;
+            padding-bottom: 16px;
+            font-weight: 700;
+            font-family: Raleway;
+        }
+
+        &6 {
+            font-family: Lato, sans-serif;
+            font-size: 16px;
+            font-weight: 400;
+            padding: 0px;
+            color: #525252;
+
+        }
+    }
+
+    .leaderImage {
+        background-color: var(--card-bg);
+
+    }
+
+    .buttonContainer {
+        margin-top: 27px;
+
+    }
+
+}
