+++ src/templates/leadership.module.css
@@ -3,47 +3,6 @@
   padding-bottom: 100px;
 }
 
-.textDecorationNone {
-  color: inherit;
-  text-decoration: none;
-  cursor: pointer;
-  align-self: center;
-}
-
-.button {
-  font-size: 16px;
-  font-weight: 600;
-
-  &:hover {
-    background: #007ebb;
-    color: #fff;
-  }
-}
-
-.heading {
-  padding-bottom: 16px;
-
-  &3 {
-    font-size: 26px;
-    padding-bottom: 16px;
-    font-weight: 700;
-  }
-
-  &6 {
-    font-family: Lato, sans-serif;
-    font-size: 17px;
-    font-weight: 400;
-    padding: 0px;
-    color: #525252;
-  }
-}
-
-.paragraph {
-  font-size: 18px;
-  font-family: Lato, sans-serif;
-  margin-bottom: 24px;
-}
-
 .imageContainer {
   display: flex; /* Establishes a flex container */
   flex-direction: row; /* Arranges items in a row */
@@ -68,4 +27,3 @@
   opacity: 1;
   }
 }
-
